+++ PACECloudSDK/AppKit/AppKit+Delegate+Slim.swift
@@ -1,0 +1,12 @@
+//
+//  AppKit+Delegate+Slim.swift
+//  PACECloudSlimSDK
+//
+//  Created by PACE Telematics GmbH.
+//
+
+import Foundation
+
+public extension AppKitDelegate {
+    func logout(completion: @escaping ((AppKit.LogoutResponse) -> Void)) {}
+}